third_party/blink/web_tests/external/wpt/html/semantics/interestfor/interestfor-invoker-descendants.tentative.html

<!DOCTYPE html>
<meta charset="utf-8" />
<link rel="author" href="mailto:masonf@chromium.org">
<link rel="help" href="https://open-ui.org/components/interest-invokers.explainer/" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="resources/invoker-utils.js"></script>

<button id=invoker interestfor=target>
  Button content
  <span id=inner-span tabindex=0>Inner content</span>
</button>
<div id=target>Target</div>
<button id=otherbutton>Other button</button>
<style>
  [interestfor] { interest-delay: 0s; }
  :interest-source { background-color: lightgreen;}
  :interest-target { background-color: lightblue;}
</style>

<script>
const invoker = document.getElementById('invoker');
const innerSpan = document.getElementById('inner-span');
const target = document.getElementById('target');
const otherbutton = document.getElementById('otherbutton');

async function showInterest(el, method) {
  if (method === 'focus') {
    await focusOn(el);
  } else {
    assert_equals(method,'hover');
    await hoverOver(el);
  }
}

['focus','hover'].forEach(method => {
  promise_test(async function (t) {
    t.add_cleanup(() => showInterest(otherbutton, method));
    const signal = t.get_signal();
    let interestCount = 0;
    let loseInterestCount = 0;
    target.addEventListener('interest',() => (++interestCount),{signal});
    target.addEventListener('loseinterest',() => (++loseInterestCount),{signal});
    await showInterest(invoker, method);
    assert_true(invoker.matches(':interest-source'),'focusing invoker should show interest');
    assert_equals(interestCount,1,'One interest event');
    interestCount = 0;
    assert_equals(loseInterestCount,0,'No loseinterest events');
    await showInterest(innerSpan, method);
    assert_true(invoker.matches(':interest-source'),'focusing inner span should keep interest');
    assert_equals(interestCount,0,'No extra interest events');
    assert_equals(loseInterestCount,0,'No loseinterest events');
    await showInterest(invoker, method);
    assert_true(invoker.matches(':interest-source'),'focusing back to outer button should keep interest');
    assert_equals(interestCount,0,'No extra interest events');
    assert_equals(loseInterestCount,0,'No loseinterest events');
    await showInterest(otherbutton, method);
    assert_false(invoker.matches(':interest-source'),'focusing outside should lose interest');
    assert_equals(interestCount,0,'No extra interest events');
    assert_equals(loseInterestCount,1,'Finally got loseinterest event');
  },`Moving focus within invoker, ${method}`);
});
</script>

<a id=outer href="#" interestfor="middle">Outer</a>
<div id="middle" popover>
  <a id="middle-link" href="#" interestfor="inner">Middle</a>
</div>
<div id="inner" popover>Inner <button id="inner-button">btn</button></div>

<script>
const outer = document.getElementById('outer');
const middle = document.getElementById('middle');
const middleLink = document.getElementById('middle-link');
const inner = document.getElementById('inner');
const innerButton = document.getElementById('inner-button');

['focus','hover'].forEach(method => {
  promise_test(async function (t) {
    t.add_cleanup(() => showInterest(otherbutton, method));
    assert_false(middle.matches(':popover-open'),'middle starts closed');
    await showInterest(outer, method);
    assert_true(outer.matches(':interest-source'),'focusing outer should show interest');
    assert_true(middle.matches(':popover-open'),'outer popover open');

    await showInterest(middleLink, method);
    assert_true(middleLink.matches(':interest-source'),'focusing middle should show interest');
    assert_true(outer.matches(':interest-source'),'outer keeps interest');
    assert_true(middle.matches(':popover-open'),'middle popover stays open');
    assert_true(inner.matches(':popover-open'),'inner popover opens');

    await showInterest(innerButton, method);
    assert_true(outer.matches(':interest-source'),'outer keeps interest');
    assert_true(middleLink.matches(':interest-source'),'middle keeps interest');
    assert_true(middle.matches(':popover-open'),'middle popover stays open');
    assert_true(inner.matches(':popover-open'),'inner popover stays open');
  },`Nested invokers, ${method}`);
});
</script>
